+++ src/builtins/netio.cc
@@ -40,7 +40,7 @@ HYBRIS_BUILTIN(hsettimeout){
 	setsockopt( data->at(0)->xint, SOL_SOCKET, SO_SNDTIMEO, &tout, sizeof(tout) );
 	setsockopt( data->at(0)->xint, SOL_SOCKET, SO_RCVTIMEO, &tout, sizeof(tout) );
 
-	return new Object(0);
+	return new Object( static_cast<long>(0) );
 }
 
 HYBRIS_BUILTIN(hconnect){
@@ -50,7 +50,7 @@ HYBRIS_BUILTIN(hconnect){
 	if( data->size() >= 2 ){
 		int sd = socket( AF_INET, SOCK_STREAM, 0 );
 		if( sd <= 0 ){
-			return new Object(-1);
+			return new Object( static_cast<long>(-1) );
 		}
 		if( data->size() == 3 ){
 			htype_assert( data->at(2), H_OT_INT );
@@ -70,10 +70,10 @@ HYBRIS_BUILTIN(hconnect){
 		}
 
 		if( connect( sd, (struct sockaddr*)&server, sizeof(server) ) != 0 ){
-			return new Object(-1);
+			return new Object( static_cast<long>(-1) );
 		}
 
-		_return = new Object( sd );
+		_return = new Object( static_cast<long>(sd) );
     }
 	else{
 		hybris_syntax_error( "function 'connect' requires at least 2 parameters (called with %d)", data->size() );
@@ -88,7 +88,7 @@ HYBRIS_BUILTIN(hserver){
 	if( data->size() >= 1 ){
 		int sd = socket( AF_INET, SOCK_STREAM, 0 );
 		if( sd <= 0 ){
-			return new Object(-1);
+			return new Object( static_cast<long>(-1) );
 		}
 		if( data->size() == 2 ){
 			htype_assert( data->at(1), H_OT_INT );
@@ -107,14 +107,14 @@ HYBRIS_BUILTIN(hserver){
 		servaddr.sin_port        = htons(port);
 
 		if ( bind( sd, (struct sockaddr *)&servaddr, sizeof(servaddr) ) < 0 ) {
-			return new Object(-1);
+			return new Object( static_cast<long>(-1) );
 		}
 
 		if ( listen( sd, 1024 ) < 0 ) {
-			return new Object(-1);
+			return new Object( static_cast<long>(-1) );
 		}
 
-		_return = new Object( sd );
+		_return = new Object( static_cast<long>(sd) );
     }
 	else{
 		hybris_syntax_error( "function 'server' requires at least 1 parameters (called with %d)", data->size() );
@@ -131,10 +131,10 @@ HYBRIS_BUILTIN(haccept){
 		    csd;
 
 		if( (csd = accept( sd, NULL, NULL) ) < 0 ) {
-			return new Object(-1);
+			return new Object( static_cast<long>(-1) );
 		}
 
-		_return = new Object( csd );
+		_return = new Object( static_cast<long>(csd) );
 	}
 	else{
 		hybris_syntax_error( "function 'accept' requires 1 parameters (called with %d)", data->size() );
@@ -179,7 +179,7 @@ HYBRIS_BUILTIN(hrecv){
 		/* handle size by type */
 		else{
 			switch( object->xtype ){
-				case H_OT_INT    : object->xsize = read = recv( sd, &object->xint,   sizeof(int), 0 ); break;
+				case H_OT_INT    : object->xsize = read = recv( sd, &object->xint,   sizeof(long), 0 ); break;
 				case H_OT_FLOAT  : object->xsize = read = recv( sd, &object->xfloat, sizeof(double), 0 ); break;
 				case H_OT_CHAR   : object->xsize = read = recv( sd, &object->xchar,  sizeof(char), 0 ); break;
 				case H_OT_STRING :
@@ -204,7 +204,7 @@ HYBRIS_BUILTIN(hrecv){
 				case H_OT_ARRAY  : hybris_generic_error( "can not directly deserialize an array type" ); break;
 			}
 		}
-        _return = new Object( static_cast<int>(read) );
+        _return = new Object( static_cast<long>(read) );
     }
 	else{
 		hybris_syntax_error( "function 'recv' requires 2 or 3 parameters (called with %d)", data->size() );
@@ -234,7 +234,7 @@ HYBRIS_BUILTIN(hsend){
 		}
 		else{
 			switch( object->xtype ){
-				case H_OT_INT    : written = send( sd, &object->xint,   sizeof(int), 0 ); break;
+				case H_OT_INT    : written = send( sd, &object->xint,   sizeof(long), 0 ); break;
 				case H_OT_FLOAT  : written = send( sd, &object->xfloat, sizeof(double), 0 ); break;
 				case H_OT_CHAR   : written = send( sd, &object->xchar,  sizeof(char), 0 ); break;
 				case H_OT_STRING : written = send( sd, object->xstring.c_str(), object->xsize, 0 ); break;
@@ -242,7 +242,7 @@ HYBRIS_BUILTIN(hsend){
 					for( i = 0; i < object->xsize; i++ ){
 						Object *element = object->xarray[i];
 						switch( element->xtype ){
-							case H_OT_INT    : written += send( sd, &element->xint,   sizeof(int), 0 ); break;
+							case H_OT_INT    : written += send( sd, &element->xint,   sizeof(long), 0 ); break;
 							case H_OT_FLOAT  : written += send( sd, &element->xfloat, sizeof(double), 0 ); break;
 							case H_OT_CHAR   : written += send( sd, &element->xchar,  sizeof(char), 0 ); break;
 							case H_OT_STRING : written += send( sd, element->xstring.c_str(), element->xsize, 0 ); break;
@@ -252,7 +252,7 @@ HYBRIS_BUILTIN(hsend){
 				break;
 			}
 		}
-        _return = new Object( static_cast<int>(written) );
+        _return = new Object( static_cast<long>(written) );
     }
 	else{
 		hybris_syntax_error( "function 'send' requires 2 or 3 parameters (called with %d)", data->size() );
@@ -264,5 +264,5 @@ HYBRIS_BUILTIN(hclose){
     if( data->size() ){
 		close( data->at(0)->xint );
     }
-    return new Object(0);
+    return new Object( static_cast<long>(0) );
 }
